switch umbra_uniform_32 and deref_ptr to size_t byte offsets
umbra_uniform_32 now takes size_t byte_off instead of int slot.
umbra_deref_ptr now takes size_t byte_off instead of int byte_off.
umbra_uniform and umbra_uniform_ptr handles use size_t off.
umbra_uniforms_size (renamed from _len) returns size_t.
umbra_draw_layout shader/coverage fields are size_t.
slug_acc_layout fields are size_t.

All byte counts and offsets are now size_t throughout the uniform
system. The .off / 4 pattern in callers is eliminated — they use
byte offsets directly (fi, fi + 4, fi + 8, ...).

+++ slides/persp.c
@@ -32,7 +32,7 @@ static void persp_draw(slide *s, int w, int h, int y0, int y1, void *buf,
     for (int i = 0; i < 4; i++) { hc[i] = s->color[i]; }
     umbra_set_f32(lay->uni, (umbra_uniform){lay->shader},   hc, 4);
     umbra_set_f32(lay->uni, (umbra_uniform){lay->coverage}, st->mat, 11);
-    umbra_set_ptr(lay->uni, (umbra_uniform_ptr){(lay->coverage + 11 * 4 + 7) & ~7},
+    umbra_set_ptr(lay->uni, (umbra_uniform_ptr){(lay->coverage + 11 * 4 + 7) & ~(size_t)7},
                   st->bitmap->data,
                   (size_t)(st->bitmap->w * st->bitmap->h * 2), 0, 0);
     size_t    pb = umbra_fmt_size(s->fmt);
